+++ WEB-INF/classes/.gitignore
@@ -1,2 +1,1 @@
-/bean/
 /todo/
