+++ src/todo/TodoList.java
@@ -1,24 +1,15 @@
 package todo;
 
 import java.io.IOException;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.naming.Context;
-import javax.naming.InitialContext;
-import javax.naming.NamingException;
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.sql.DataSource;
 
 import bean.Member;
 
@@ -27,82 +18,34 @@ public class TodoList extends HttpServlet {
 
 	static final boolean useConnectionPool = false;
 
+	/**
+	 * "/todo"へのGETリクエストをdoPost()へ渡します
+	 * @param request リクエスト
+	 * @param response レスポンス
+	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		// 常にPOSTへ渡します。
 		doPost(request, response);
 	}
 
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-			request.setCharacterEncoding("UTF8");
-
-			// 呼び出し先Jspに渡すデータセット
-			request.setAttribute("list", getDbData());
-
-			// result.jsp にページ遷移
-			RequestDispatcher dispatch = request.getRequestDispatcher("/WebContent/JSP/list.jsp");
-			dispatch.forward(request, response);
-	}
-
-	private List<Member> getDbData() {		
-		List<Member> list = new ArrayList<Member>();
-		Connection con = null;
-		Statement stmt = null;
-		ResultSet rs = null;
-
-		try {
-			con = getConnection();
-			stmt = con.createStatement();
-			rs = stmt.executeQuery("select id, task from todo where delflg = 0 order by createtime desc");
-
-			while (rs.next()) {
-				String id = rs.getString("id");
-				String task = rs.getString("task");
-				list.add(new Member(id,task));
-			}
-		} catch (SQLException e) {
-			e.printStackTrace();
-//			return "エラー発生(SQLException)";
-		} catch (NamingException e) {
-			e.printStackTrace();
-//			return "エラー発生(NamingException)";
-		} catch (ClassNotFoundException e) {
-			e.printStackTrace();
-//			return "エラー発生(ClassNotFoundException)";
-		} finally {
-			try { if (rs   != null) rs.close();   } catch (Exception e) { e.printStackTrace(); }
-			try { if (stmt != null) stmt.close(); } catch (Exception e) { e.printStackTrace(); }
-			try { if (con  != null) con.close();  } catch (Exception e) { e.printStackTrace(); }
-		}
-		return list;
-	}
-
 	/**
-	 * DBとの接続を取得します。
-	 * @return DBとのConnection
-	 * @throws NamingException
-	 * @throws SQLException
-	 * @throws ClassNotFoundException
+	 * DBからデータを取得しTODO一覧画面を教示します
+	 * @param request リクエスト
+	 * @param response レスポンス
 	 */
-	Connection getConnection() throws NamingException, SQLException, ClassNotFoundException {
-		final String dataSourceName = "java:comp/env/jdbc/postgresql";
-
-		final String dbUrl = "jdbc:postgresql://localhost:5432/postgres";
-		final String dbUser = "postgres";
-		final String dbPass = "postgres";
-
-		Connection con = null;
-
-		if (useConnectionPool) {
-			System.out.println("コネクションプールよりDB接続を取得します。");
-			Context context = new InitialContext();
-			DataSource ds = (DataSource) context.lookup(dataSourceName);
-			con = ds.getConnection();
-		} else {
-			System.out.println("DBへ直接接続を行います。");
-			Class.forName("org.postgresql.Driver");
-			con = DriverManager.getConnection(dbUrl, dbUser, dbPass);
-		}
-
-		return con;
+	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+			GetData getdata = new GetData();
+			List<Member> list = new ArrayList<Member>();
+			
+			request.setCharacterEncoding("UTF8");
+			
+			if (getdata.getDbData(list)) {
+				request.setAttribute("list", list);
+				RequestDispatcher dispatch = request.getRequestDispatcher("/WebContent/JSP/list.jsp");
+				dispatch.forward(request, response);
+				return;
+			} else {
+				RequestDispatcher dispatch = request.getRequestDispatcher("WebContent/JSP/error.jsp");
+				dispatch.forward(request, response);
+			}
 	}
 }
