+++ work/org/apache/jsp/WebContent/JSP/add_jsp.java
@@ -1,7 +1,7 @@
 /*
  * Generated by the Jasper component of Apache Tomcat
  * Version: Apache Tomcat/8.5.73
- * Generated at: 2022-01-27 07:53:22 UTC
+ * Generated at: 2022-01-28 02:51:12 UTC
  * Note: The last modified time of this file was set to
  *       the last modified time of the source file after
  *       generation to assist with modification tracking.
@@ -123,26 +123,30 @@ public final class add_jsp extends org.apache.jasper.runtime.HttpJspBase
       out.write("	<title>TODO入力画面</title>\n");
       out.write("</head>\n");
       out.write("<body>\n");
+      out.write("	<div class=\"container-xxl\">\n");
+      out.write("		<p class=\"title\">TODO入力画面</p>\n");
       out.write("	");
 
         if(request.getAttribute("error_msg")!=null){
       out.write("\n");
-      out.write("                ");
+      out.write("                <p class=\"error-msg\">");
       out.print(request.getAttribute("error_msg") );
-      out.write("\n");
+      out.write("</p>\n");
       out.write("        ");
 }
 	
       out.write("\n");
-      out.write("	<div class=\"container-xxl\">\n");
-      out.write("		<p class=\"title\">TODO入力画面</p>\n");
       out.write("		<form action='/create' method='post' onSubmit=\"return check()\">\n");
       out.write("			<div class=\"row\">\n");
       out.write("				<label class=\"col\">新規TODO</label>\n");
       out.write("			</div>\n");
       out.write("			<div class=\"row\">\n");
       out.write("				<div class=\"col-10\">\n");
-      out.write("					<textarea type=\"text\" id=\"new_todo\" name=\"new_todo\" placeholder='新しいTODOを入力してください。'></textarea>\n");
+      out.write("					<textarea type=\"text\" id=\"new_todo\" name=\"new_todo\" placeholder='新しいTODOを入力してください。'>");
+ if(request.getAttribute("value")!=null){ 
+      out.print(request.getAttribute("value"));
+ } 
+      out.write("</textarea>\n");
       out.write("				</div>		\n");
       out.write("			</div>\n");
       out.write("			<div class=\"row\">\n");
