+++ work/org/apache/jsp/WebContent/JSP/list_jsp.java
@@ -1,7 +1,7 @@
 /*
  * Generated by the Jasper component of Apache Tomcat
  * Version: Apache Tomcat/8.5.73
- * Generated at: 2022-01-27 07:52:15 UTC
+ * Generated at: 2022-01-27 23:47:24 UTC
  * Note: The last modified time of this file was set to
  *       the last modified time of the source file after
  *       generation to assist with modification tracking.
@@ -11,9 +11,6 @@ package org.apache.jsp.WebContent.JSP;
 import javax.servlet.*;
 import javax.servlet.http.*;
 import javax.servlet.jsp.*;
-import bean.Member;
-import java.util.List;
-import java.util.ArrayList;
 
 public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
     implements org.apache.jasper.runtime.JspSourceDependent,
@@ -39,10 +36,7 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
     _jspx_imports_packages.add("javax.servlet");
     _jspx_imports_packages.add("javax.servlet.http");
     _jspx_imports_packages.add("javax.servlet.jsp");
-    _jspx_imports_classes = new java.util.HashSet<>();
-    _jspx_imports_classes.add("java.util.List");
-    _jspx_imports_classes.add("bean.Member");
-    _jspx_imports_classes.add("java.util.ArrayList");
+    _jspx_imports_classes = null;
   }
 
   private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems;
@@ -127,9 +121,6 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
 
       out.write("\n");
       out.write("\n");
-      out.write("\n");
-      out.write("\n");
-      out.write("\n");
       out.write("<!DOCTYPE html>\n");
       out.write("<html>\n");
       out.write("<head>\n");
@@ -181,8 +172,7 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
       out.write("		}\n");
       out.write("	</script>\n");
       out.write("</body>\n");
-      out.write("</html>\n");
-      out.write("\n");
+      out.write("</html>");
     } catch (java.lang.Throwable t) {
       if (!(t instanceof javax.servlet.jsp.SkipPageException)){
         out = _jspx_out;
@@ -212,9 +202,9 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
     try {
       _jspx_th_c_005fforEach_005f0.setPageContext(_jspx_page_context);
       _jspx_th_c_005fforEach_005f0.setParent(null);
-      // /WebContent/JSP/list.jsp(26,5) name = items type = javax.el.ValueExpression reqTime = true required = false fragment = false deferredValue = true expectedTypeName = java.lang.Object deferredMethod = false methodSignature = null
-      _jspx_th_c_005fforEach_005f0.setItems(new org.apache.jasper.el.JspValueExpression("/WebContent/JSP/list.jsp(26,5) '${list}'",_jsp_getExpressionFactory().createValueExpression(_jspx_page_context.getELContext(),"${list}",java.lang.Object.class)).getValue(_jspx_page_context.getELContext()));
-      // /WebContent/JSP/list.jsp(26,5) name = var type = java.lang.String reqTime = false required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
+      // /WebContent/JSP/list.jsp(23,5) name = items type = javax.el.ValueExpression reqTime = true required = false fragment = false deferredValue = true expectedTypeName = java.lang.Object deferredMethod = false methodSignature = null
+      _jspx_th_c_005fforEach_005f0.setItems(new org.apache.jasper.el.JspValueExpression("/WebContent/JSP/list.jsp(23,5) '${list}'",_jsp_getExpressionFactory().createValueExpression(_jspx_page_context.getELContext(),"${list}",java.lang.Object.class)).getValue(_jspx_page_context.getELContext()));
+      // /WebContent/JSP/list.jsp(23,5) name = var type = java.lang.String reqTime = false required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
       _jspx_th_c_005fforEach_005f0.setVar("item");
       int[] _jspx_push_body_count_c_005fforEach_005f0 = new int[] { 0 };
       try {
@@ -272,7 +262,7 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
     try {
       _jspx_th_c_005fout_005f0.setPageContext(_jspx_page_context);
       _jspx_th_c_005fout_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fforEach_005f0);
-      // /WebContent/JSP/list.jsp(28,26) name = value type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
+      // /WebContent/JSP/list.jsp(25,26) name = value type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
       _jspx_th_c_005fout_005f0.setValue((java.lang.Object) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${item.task}", java.lang.Object.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
       int _jspx_eval_c_005fout_005f0 = _jspx_th_c_005fout_005f0.doStartTag();
       if (_jspx_th_c_005fout_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
